+++ biahub/cli/estimate_registration.py
@@ -19,12 +19,6 @@ from biahub.analysis.AnalysisSettings import (
     StabilizationSettings,
 )
 from biahub.analysis.analyze_psf import detect_peaks
-from biahub.analysis.register import (
-    convert_transform_to_ants,
-    convert_transform_to_numpy,
-    get_3D_rescaling_matrix,
-    get_3D_rotation_matrix,
-)
 from biahub.cli.parsing import (
     config_filepath,
     num_processes,
@@ -33,6 +27,12 @@ from biahub.cli.parsing import (
     target_position_dirpaths,
 )
 from biahub.cli.utils import _check_nan_n_zeros, model_to_yaml, yaml_to_model
+from biahub.register import (
+    convert_transform_to_ants,
+    convert_transform_to_numpy,
+    get_3D_rescaling_matrix,
+    get_3D_rotation_matrix,
+)
 
 # TODO: see if at some point these globals should be hidden or exposed.
 NA_DETECTION_SOURCE = 1.35
