+++ biahub/cli/main.py
@@ -4,7 +4,6 @@ from biahub.cli.characterize_psf import characterize_psf
 from biahub.cli.estimate_psf import estimate_psf
 from biahub.cli.estimate_registration import estimate_registration
 from biahub.cli.optimize_registration import optimize_registration
-from biahub.cli.register import register
 from biahub.cli.stabilize import stabilize
 from biahub.concatenate import concatenate_cli
 from biahub.deconvolve import deconvolve_cli
@@ -13,6 +12,7 @@ from biahub.estimate_bleaching import estimate_bleaching_cli
 from biahub.estimate_deskew import estimate_deskew_cli
 from biahub.estimate_stabilization import estimate_stabilization_cli
 from biahub.estimate_stitch import estimate_stitch_cli
+from biahub.register import register_cli
 from biahub.segment import segment_cli
 from biahub.stitch import stitch_cli
 
@@ -35,7 +35,7 @@ cli.add_command(estimate_deskew_cli)
 cli.add_command(deskew_cli)
 cli.add_command(estimate_registration)
 cli.add_command(optimize_registration)
-cli.add_command(register)
+cli.add_command(register_cli)
 cli.add_command(estimate_stitch_cli)
 cli.add_command(stitch_cli)
 cli.add_command(concatenate_cli)
