+++ biahub/tests/test_analysis/test_affine.py
@@ -2,7 +2,7 @@ import ants
 import numpy as np
 import pytest
 
-from biahub.analysis.register import (
+from biahub.register import (
     apply_affine_transform,
     convert_transform_to_ants,
     convert_transform_to_numpy,
