+++ biahub/tests/test_cli/test_register_cli.py
@@ -4,7 +4,7 @@ from click.testing import CliRunner
 from numpy import testing
 
 from biahub.cli.main import cli
-from biahub.cli.register import rescale_voxel_size
+from biahub.register import rescale_voxel_size
 
 
 def test_register_cli(tmp_path, example_plate, example_plate_2, example_register_settings):
